+++ lib/believer/command.rb
@@ -23,9 +23,9 @@ module Believer
         cql = to_cql
         begin
           start = Time.now
-          puts "Executing #{cql}"
+          #puts "Executing #{cql}"
           res = connection.execute(cql)
-          #Rails.logger.debug "#{name} #{sprintf "%.3f", (Time.now - start)*1000.0} ms: #{cql}"
+          puts "#{name} #{sprintf "%.3f", (Time.now - start)*1000.0} ms: #{cql}"
           return res
         rescue Cql::Protocol::DecodingError => e
           # Decoding errors tend to #$%# up the connection, resulting in no more activity, so a reconnect is performed here.