+++ lib/believer/querying.rb
@@ -1,6 +1,6 @@
 module Believer
   module Querying
-    delegate :select, :where, :order, :limit, :to => :scoped
+    delegate :select, :where, :order, :limit, :exists?, :to => :scoped
 
   end
 end